Added option to RenderPartialAsync when you don't need a layout page around the partial view.
