+++ examples/Mvc/ExampleWebApp.Net6_0/Views/Home/Index.cshtml
@@ -6,3 +6,10 @@
     <h1 class="display-4">Welcome</h1>
     <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
 </div>
+
+<ul>
+    <li><a href="/home/RenderRcl">Render content using templating engine</a></li>
+    <li><a href="/home/RenderPartialTest">Render partial content using templating engine</a></li>
+</ul>
+
+
