+++ examples/Mvc/ExampleWebApp.Net6_0/Views/Home/_partialTest.cshtml
@@ -1,0 +1,8 @@
+﻿@*
+    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+*@
+@{
+    var testval = Model as string;
+}
+<h1>This is a partial page</h1>
+<h2>@testval</h2>
