+++ examples/Templates/ExampleAppRazorTemplates/Views/_examplePartialView.cshtml
@@ -1,0 +1,10 @@
+﻿@*
+    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+*@
+@{
+    var testvalue = Model as string;
+}
+<h1>This is a partial view</h1>
+<p>This is from a model:
+    <strong>@testvalue</strong>
+</p>
